requirements.txt
Django==5.2
psycopg2-binary==2.9.10
django-storages==1.14.6
boto3==1.37.37
djangorestframework==3.16.0
django-debug-toolbar==6.1.0
